+++ main/filesystem_utils.h
@@ -8,5 +8,6 @@ bool      fs_utils_is_directory(const char* path);
 bool      fs_utils_is_file(const char* path);
 esp_err_t fs_utils_remove(const char* path);
 esp_err_t fs_utils_copy_recursive(const char* src, const char* dst);
+uint64_t  fs_utils_get_directory_size(const char* path);
 size_t    fs_utils_get_file_size(FILE* fd);
 uint8_t*  fs_utils_load_file_to_ram(FILE* fd);
